Make SelectorState detect force substitutions
Prior to this change, a `ResolutionStrategy.force` based force was not
handled the same as setting `force = true` on a dependency.
This caused virtual platforms alignment to not properly detect there
were competing force in action.
With this change, a `SelectorState` is marked as force also for a forced
substitution.

Fixes #6817